+++ kotti_newsitem/tests/test_configure.py
@@ -14,7 +14,7 @@ def test_kotti_configure():
     kotti_configure(settings)
 
     assert settings['pyramid.includes'] == ' kotti_newsitem'
-    assert settings['kotti.available_types'] == ' kotti_newsitem.resources.NewsItem kotti_newsitem.resources.NewsPage'
+    assert settings['kotti.available_types'] == ' kotti_newsitem.resources.NewsItem'
 
     assert 'kotti_newsitem.widget.slot' not in settings
     assert settings['kotti_newsitem.widget.num_news'] == 5