+++ kotti_newsitem/tests/test_views.py
@@ -2,12 +2,14 @@ from datetime import date
 from random import randint
 
 from kotti.resources import get_root
+from kotti.resources import Document
+from kotti.views.view import view as kotti_view
+
+from kotti.views.edit.default_views import DefaultViewSelection
 
 from kotti_newsitem.resources import NewsItem
-from kotti_newsitem.resources import NewsPage
 from kotti_newsitem.views import NewsItemListViews
 from kotti_newsitem.views import NewsItemView
-from kotti_newsitem.views import NewsPageView
 
 
 def test_newsitem_view(db_session, dummy_request):
@@ -62,46 +64,6 @@ def test_news_item_list_views(db_session, dummy_request):
         assert n.publish_date <= d
         d = n.publish_date
 
-
-def test_newspage_view(db_session, dummy_request):
-
-    root = get_root()
-    content = NewsPage()
-    root['content'] = content
-
     dummy_request.registry.settings['kotti_newsitem.num_news'] = 4
-    view = NewsPageView(root['content'], dummy_request)
 
-    assert view.news_page()['items'] == []
-
-
-def test_news_page_view(db_session, dummy_request):
-
-    root = get_root()
-    d = date.today()
-
-    # Add some news items
-    for i in range(10):
-
-        n = NewsItem(
-            title=u"News Item %s Title" % (i + 1),
-            description=u"News Item %s description" % (i + 1),
-            )
-
-        if i == 9:
-            # Set date to future for the last news item
-            # this will fail in leap years on february 29th, which is acceptable
-            # as this is only a test.
-            n.publish_date = date(d.year + 1, d.month, d.day)
-        else:
-            # Set date to random year in the past
-            n.publish_date = date(d.year - randint(0, 100), d.month, d.day)
-
-        root["news_item_%s" % (i + 1)] = n
-        db_session.add(n)
-        db_session.flush()
-
-    dummy_request.registry.settings['kotti_newsitem.num_news'] = 4
-    view = NewsPageView(root, dummy_request)
-
-    assert len(view.news_page()['items']) == 4
+    assert len(view.news_listing()['items']) == 4